change colorscheme stuff (transparent bg)
Change to transparent background (technically "none'), so it now seems to depend on the
background colour the terminal is set to. Couldn't get something to work
for crosshair but got some cool insights into how to change settings for
cursor crosshair.
Also changed colorscheme to wombat256i, with i for improved. It's
different and not always better (eg. search highlighting is ugly) so I
might change back.